+++ crates/goose-cli/src/scenario_tests/scenarios.rs
@@ -1,11 +1,11 @@
 #[cfg(test)]
 mod tests {
-    use crate::scenario_tests::run_multi_provider_scenario;
+    use crate::scenario_tests::scenario_runners::run_scenario;
     use anyhow::Result;
 
     #[tokio::test]
     async fn test_what_is_your_name() -> Result<()> {
-        run_multi_provider_scenario("what_is_your_name", &["what is your name"], |result| {
+        run_scenario("what_is_your_name", &["what is your name"], |result| {
             assert!(result
                 .message_contents()
                 .iter()
